Fix Gdip_FastImageSearch() memory leak
The memory leak is gone. Now, each static bitmap adds +4 MB to the
script's RAM only while it's saved/locked. This is expected, as +8MB is
observed when both the haystack and the needle are locked. At the moment
they are released and disposed, this same amount of memory is freed.